stock-predictor/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph>=0.0.62
langchain>=0.1.0
langchain-openai>=0.0.2
yfinance==0.2.28
pandas==2.1.4
numpy==1.25.2
ta==0.10.2
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
pyarrow==14.0.1
alpha-vantage==2.3.1
requests==2.31.0
aiohttp==3.9.1
ollama==0.5.3
